+++ src/messages/components/NotificationToggle.vue
@@ -6,6 +6,9 @@
     :color="inToolbar ? 'white' : selected.color"
     :size="size"
   >
+    <QTooltip>
+      {{ $t('NOTIFICATION_TOGGLE.HEADER') }}
+    </QTooltip>
     <QIcon
       :name="selected.icon"
     />
@@ -52,6 +55,7 @@
 <script>
 import {
   QPopover,
+  QTooltip,
   QBtn,
   QIcon,
   QList,
@@ -65,6 +69,7 @@ import {
 export default {
   components: {
     QPopover,
+    QTooltip,
     QBtn,
     QIcon,
     QList,