+++ src/pickups/components/PickupList.vue
@@ -1,8 +1,22 @@
 <template>
   <div>
+    <QCard
+      v-if="filter && pickups.length > 0"
+      flat
+      class="generic-padding row no-wrap items-center justify-between"
+    >
+      <QSelect
+        v-model="slots"
+        :options="slotsOptions"
+        hide-underline
+      />
+      <div class="q-caption q-ml-xs">
+        {{ filteredPickups.length }} / {{ pickups.length }}
+      </div>
+    </QCard>
     <KSpinner v-show="pending" />
     <PickupItem
-      v-for="pickup in pickups"
+      v-for="pickup in filteredPickups"
       :key="pickup.id"
       :pickup="pickup"
       :place-link="placeLink"
@@ -16,12 +30,25 @@
 <script>
 import PickupItem from './PickupItem'
 import KSpinner from '@/utils/components/KSpinner'
+import bindRoute from '@/utils/mixins/bindRoute'
+
+import {
+  QSelect,
+  QCard,
+} from 'quasar'
 
 export default {
   components: {
     PickupItem,
     KSpinner,
+    QSelect,
+    QCard,
   },
+  mixins: [
+    bindRoute({
+      slots: 'all',
+    }),
+  ],
   props: {
     pickups: {
       type: Array,
@@ -34,6 +61,50 @@ export default {
     pending: {
       type: Boolean,
       default: false,
+    },
+    filter: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  watch: {
+    slots (val, old) {
+      // keep selection valid, revert to old value or default
+      const options = this.slotsOptions.map(o => o.value)
+      if (!options.includes(val)) {
+        if (options.includes(old)) {
+          this.slots = old
+        }
+        else {
+          this.slots = options[0]
+        }
+      }
+    },
+  },
+  data () {
+    return {
+      slotsOptions: [
+        {
+          label: this.$t('PICKUPLIST.FILTER.ALL'),
+          value: 'all',
+        },
+        {
+          label: this.$t('PICKUPLIST.FILTER.FREE'),
+          value: 'free',
+        },
+        {
+          label: this.$t('PICKUPLIST.FILTER.EMPTY'),
+          value: 'empty',
+        },
+      ],
+    }
+  },
+  computed: {
+    filteredPickups () {
+      if (!this.pickups) return []
+      if (this.slots === 'free') return this.pickups.filter(e => !e.isFull)
+      if (this.slots === 'empty') return this.pickups.filter(e => e.isEmpty)
+      return this.pickups
     },
   },
 }