+++ CHANGELOG.md
@@ -25,12 +25,16 @@ Please document your changes in this format:
 - Add ability to specify pickups with end times #709 @nicksellen
 - Add store wall #986 @tiltec
 - Add feature to mark stores as favorite #986 @tiltec
+- Add slots filter to group pickup list @tiltec
 
 ### Changed
 - Allow non-collectors to read and write in pickup chat @tiltec
 - Pickups from non-favorited stores won't show up in the "available pickup" in the group wall and don't send emails about upcoming pickups @tiltec
 - Can now unsubscribe from conversations (except private conversation and threads) @tiltec
 - Show store name instead of "Current store" in sidenav @tiltec
+
+### Fixed
+- Ended conversations weren't closed if they had no messages @tiltec
 
 ## [6.5.0] - 2019-02-03
 ### Added